Adapted changes to backend modifications

+++ src/components/Navbar/navbar.jsx
@@ -3,7 +3,10 @@ import "./navbar.css";
 import Logo from '../../assets/images/logo.png'
 import {Link, useNavigate} from "react-router-dom";
 import {useLocation} from "react-router-dom";
-import {fetchUserData} from "../../api/authenticationService";
+import {fetchUserData, userLogout, userRefreshToken} from "../../api/authenticationService";
+import {jwtDecode} from "jwt-decode";
+import {authenticate, authSuccess} from "../../redux/authActions";
+import {connect} from "react-redux";
 
 
 const Navbar = (props) => {
@@ -13,6 +16,8 @@ const Navbar = (props) => {
     const [data, setData] = useState({});
     const [isLoggedIn, setIsLoggedIn] = useState(false);
     const [searchTerm, setSearchTerm] = useState('');
+    const token = localStorage.getItem('ACCESS_TOKEN');
+
 
     const handleSearchSubmit = (event) => {
         event.preventDefault();
@@ -20,28 +25,58 @@ const Navbar = (props) => {
     };
 
     React.useEffect(() => {
-        fetchUserData().then((response) => {
-            setData(response.data);
-        }).catch(() => {
-            localStorage.clear();
-            props.history.push('/');
+        if (token) {
+            fetchUserData().then((response) => {
+                setData(response.data);
+            }).catch(() => {
+                localStorage.clear();
+                // props.history.push('/');
 
-        })
+            })
+        }
     }, [])
 
-    useEffect(() => {
-        const token = localStorage.getItem('ACCESS_TOKEN');
-        if (token) {
-            setIsLoggedIn(true);
-        }
 
-    }, [setIsLoggedIn]);
+    useEffect(() => {
+        const checkTokenValidity = () => {
+            if (token) {
+                const decodedToken = jwtDecode(token);
+                const expirationDate = new Date(decodedToken.exp * 1000 - 10000);
+                console.log("Expiration Date:", expirationDate);
+                console.log("Current Date:", new Date());
+
+                if (expirationDate <= new Date()) {
+                    console.log("Refreshing token...");
+                    userRefreshToken().then((res) => {
+                        if (res.status === 200) {
+                            localStorage.clear()
+                            props.setUser(res.data);
+                            console.log("Token refreshed successfully.", res.data);
+                            window.location.reload()
+                        }
+                    });
+                } else {
+                    console.log("Token is still valid");
+                    setIsLoggedIn(true);
+                }
+            }
+            setTimeout(checkTokenValidity, 20000);
+        };
+        checkTokenValidity();
+        return () => clearTimeout(checkTokenValidity);
+    }, []);
 
     const logout = () => {
-        localStorage.clear();
-        setIsLoggedIn(false);
-        navigate('/');
+        userLogout().then((res) => {
+            console.log(res)
+            if (res.status === 200) {
+                localStorage.clear();
+                setIsLoggedIn(false);
+                navigate('/');
+            }
+        }).catch(err => console.log(err))
     }
+
 
     if (pathname === "/login" || pathname === "/join") {
         return null;
@@ -105,4 +140,21 @@ const Navbar = (props) => {
     );
 };
 
-export default Navbar;
+const mapStateToProps = ({auth}) => {
+    console.log("state ", auth)
+    return {
+        loading: auth.loading,
+        error: auth.error
+    }
+}
+const mapDispatchToProps = (dispatch) => {
+    return {
+        authenticate: () => dispatch(authenticate()),
+        setUser: (content) => {
+            console.log("Content:", content);
+            dispatch(authSuccess(content))
+        },
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Navbar);
